CORE-396 time synching between ec2 instances
Not perfect but you need some kind of GPS satellite stuff to get it accurate on
a millisecond scale. In the meanwhile, I assume time errors are within 10 ms
range so that if a time difference is less than zero I just zero it. If the
difference is less than -10, an errors is thrown. I may need to increase that
limit though.